Update README.md for PE Master setup
This commit updates the README to include a quick start guide which
builds a Puppet Enterprise master using the gce_instance Puppet type.
It also provides instructions for using a newly provided manifest
to automate some of the demo configuration necessary, without
requiring any PE knowledge on behalf of the user.